feat: expand number exts to other number types + add isNullOrZero

+++ kotlin-utils/src/main/kotlin/xyz/lbres/kotlinutils/number/ByteExt.kt
@@ -5,28 +5,28 @@ import xyz.lbres.kotlinutils.utils.simpleIf
 /**
  * Returns this number if not zero, or the result of calling [getDefaultValue] if it is.
  *
- * @param getDefaultValue () -> [Int]
- * @return [Int] the current value, or the default
+ * @param getDefaultValue () -> [Byte]
+ * @return [Byte] the current value, or the default
  */
-fun Int.ifZero(getDefaultValue: () -> Int): Int = simpleIf(isZero(), { getDefaultValue() }, { this })
+fun Byte.ifZero(getDefaultValue: () -> Byte): Byte = simpleIf(isZero(), { getDefaultValue() }, { this })
 
 /**
  * Unary check to determine if value is zero
  *
  * @return [Boolean]: true if value is zero, false otherwise
  */
-fun Int.isZero(): Boolean = equals(0)
+fun Byte.isZero(): Boolean = equals(0.toByte())
 
 /**
  * Unary check to determine if value is negative
  *
  * @return [Boolean]: true if value is less than zero, false otherwise
  */
-fun Int.isNegative(): Boolean = this < 0
+fun Byte.isNegative(): Boolean = this < 0.toByte()
 
 /**
  * Returns true if value is null or zero, or false otherwise
  *
  * @return [Boolean]
  */
-fun Int?.isNullOrZero(): Boolean = this == null || isZero()
+fun Byte?.isNullOrZero(): Boolean = this == null || isZero()
